requirements.txt
matplotlib==2.0.2
numpy==1.13.0
pyquaternion==0.9.0
pyserial==3.3
scikit-learn==0.18.2
scipy==0.19.1
sklearn==0.0
smbus